useMap renamed to use-map
